requirements.txt
streamlit>=1.24.0
numpy<2.0.0
pandas>=1.5.0,<2.0.0
seaborn==0.12.0
matplotlib==3.7.0
pillow>=9.4.0
python-chess>=1.10.0
requests>=2.28.0
scikit-learn>=1.2.0
statsmodels>=0.13.0
setuptools>=65.5.1
mord==0.7 